Porting in chrome/ (add 6 files to Linux SCons build)
Review URL: http://codereview.chromium.org/17052

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7725 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/renderer_host/download_resource_handler.h
@@ -66,7 +66,7 @@ class DownloadResourceHandler : public ResourceHandler {
   base::OneShotTimer<DownloadResourceHandler> pause_timer_;
 
   static const int kReadBufSize = 32768;  // bytes
-  static const int kLoadsToWrite = 100;  // number of data buffers queued
+  static const size_t kLoadsToWrite = 100;  // number of data buffers queued
   static const int kThrottleTimeMs = 200;  // milliseconds
 
   DISALLOW_COPY_AND_ASSIGN(DownloadResourceHandler);
